feat(FieldsetCommunity): hook 추가 및 리팩토링 수행
