+++ src/components/Community/WriteBackground/WriteBackground.jsx
@@ -1,23 +1,10 @@
 /** @jsxImportSource @emotion/react */
-import { css } from '@emotion/react';
 import WriteContainer from '../WriteContainer/WriteContainer';
-
-const colors = {
-  lightBlue: '#5C8AFF',
-  darkBlue: '#37415B',
-};
-
-const style = color => css`
-  background-color: ${colors[color]};
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-`;
+import { backgroundStyle } from './WriteBackground.style';
 
 export default function WriteBackground({ color }) {
   return (
-    <div css={style(color)}>
+    <div css={backgroundStyle(color)}>
       <WriteContainer />
     </div>
   );
