+++ src/components/Community/WriteContainer/WriteContainer.jsx
@@ -1,36 +1,6 @@
 /** @jsxImportSource @emotion/react */
-import { css } from '@emotion/react';
 import FieldsetCommunity from '../FieldsetCommunity/FiledsetCommunity';
-
-// height는 content만큼 늘어나도록 설정
-const parentContainer = css`
-  width: 90vw;
-  overflow: hidden;
-  height: content;
-  background-color: #f8fafb;
-  border-radius: 20px;
-  margin-top: 3em;
-  margin-bottom: 10rem;
-
-  display: flex;
-  justify-content: space-around;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const childContainer = css`
-  width: 90%;
-  height: content;
-  overflow: hidden;
-
-  background-color: #ffffff;
-  border-radius: 20px;
-  border: 1px solid #dfdfdf;
-
-  display: flex;
-  flex-driection: column;
-  align-items: center;
-`;
+import { parentContainer, childContainer } from './WriteContainer.style';
 
 export default function WriteContainer() {
   return (
